+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added support for "journal entry" type account activities
+- Introduced `--begin` argument to `account activity get` subcommand
 - Bumped minimum supported Rust version to `1.63`
 - Bumped `apca` dependency to `0.28.0`